% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmetadata-package.R
\docType{package}
\name{kmetadata-package}
\alias{kmetadata}
\alias{kmetadata-package}
\title{kmetadata: kmetadata - R package for managing metadata}
\description{
The kmetadata package provides tools for managing and manipulating metadata in R.
}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
